requirements.txt
fastapi
uvicorn
pymongo
sentence-transformers
scikit-learn
pdfminer.six
requests
python-docx
google-generativeai